+++ front/src/template.rs
@@ -1,12 +1,10 @@
-use std::net::SocketAddr;
-
 use axum::http::request::Parts;
 use axum::{async_trait, extract::FromRequestParts};
 use front_components::Ref;
 use maud::{html, Markup, PreEscaped, DOCTYPE};
 use strum::{EnumIter, IntoEnumIterator};
 
-use crate::{AppState, Padded};
+use crate::{AppState, Ip, Padded};
 
 #[derive(Debug, Clone, Copy)]
 pub enum ContentMode {
@@ -16,9 +14,11 @@ pub enum ContentMode {
 
 impl Template {
     pub async fn new(parts: &Parts, state: &AppState) -> Self {
+        let db = state.surrealdb.get().await.unwrap();
+
         Template {
-            ips: state.registered_ips.read().await.clone(),
-            selected_ip: *state.selected_ip.read().await,
+            ips: db.select("ips").await.unwrap(),
+            selected_ip: state.selected_ip.read().await.clone(),
             title: format!("ADAM - {}", parts.uri.path()),
             mode: if parts.headers.get("HX-Request").is_some() {
                 ContentMode::Embedded
@@ -32,8 +32,8 @@ impl Template {
 pub struct Template {
     title: String,
     mode: ContentMode,
-    ips: Vec<SocketAddr>,
-    selected_ip: Option<SocketAddr>,
+    ips: Vec<Ip>,
+    selected_ip: Option<Ip>,
 }
 
 #[allow(dead_code)]
@@ -121,8 +121,8 @@ async fn Template(
     title: &str,
     mode: ContentMode,
     content: Markup,
-    ips: Vec<SocketAddr>,
-    selected_ip: Option<SocketAddr>,
+    ips: Vec<Ip>,
+    selected_ip: Option<Ip>,
 ) -> Markup {
     if let ContentMode::Embedded = mode {
         return html! {
@@ -140,7 +140,6 @@ async fn Template(
                 script defer src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js" {}
                 script src="https://unpkg.com/htmx.org" {}
                 script src="https://cdn.tailwindcss.com" {}
-                script src="https://cdn.jsdelivr.net/npm/alpinejs@3.x.x/dist/cdn.min.js" defer {}
                 script {
                     "
                         function toggleDarkMode() {
@@ -206,7 +205,7 @@ async fn Template(
                     }
 
                     // IP Selector Dropdown
-                    @if let Some(selected_ip) = selected_ip {
+                    @if let Some(selected_ip) = &selected_ip {
                         div.flex.flex-row.items-center."space-x-4" {
                             form {
                                 label for="ip-select" { "Select IP: " }
@@ -215,10 +214,12 @@ async fn Template(
                                     id="ip-select"
                                 {
                                     @for ip in &ips {
-                                        @if ip == &selected_ip {
-                                            option value=(ip) selected { (ip) }
+                                        @if ip == selected_ip {
+                                            option value=(ip.id.as_ref().unwrap().id.to_string()) selected
+                                            { (ip.name) " (" (ip.socket) ")" }
                                         } @else {
-                                            option value=(ip) { (ip) }
+                                            option value=(ip.id.as_ref().unwrap().id.to_string())
+                                            { (ip.name) " (" (ip.socket) ")" }
                                         }
                                     }
                                 }
@@ -232,17 +233,14 @@ async fn Template(
                 (Footer())
             }
 
-            @if selected_ip.is_some() {
+            @if selected_ip.as_ref().is_some() {
                 script type="text/javascript" {
                     (PreEscaped(r#"
                     document.getElementById('ip-select').addEventListener('change', async function() {
                         const selectedIp = this.value;
-                        await fetch('/api/ips/selected', {
-                            method: 'POST',
-                            headers: { 'Content-Type': 'application/json' },
-                            body: JSON.stringify({ ip: selectedIp })
+                        await fetch(`/ips/${selectedIp}`, {
+                            method: 'PATCH',
                         });
-                        // Optionally, refresh content that depends on the selected IP ?
                     });
                     "#))
                 }
